More flexible specification of payoffs for outcomes.
This implements a more sophisticated approach to setting payoffs
associated with outcomes.  Previously, the representation was quite
militant in insisting only int, Decimal, or Fraction were acceptable
payoffs; in particular, floats (or any numpy type) were not.

Now, if the type of the specified payoff is not int, Decimal, or Fraction,
the payoff setter attempts to create a decimal or rational number
based on the string representation of the passed payoff.  This does
so in a way that passing a float should not result in unexpected
precision errors, as it is done via the decimal.Decimal() contructure.

In addition, this means that passing strings directly to the outcome
payoffs works - so values like "1/2" or "18.1" will do the right
thing, without calling code having to construct a Fraction or Decimal
itself.